Initial sketch of separate login strategies
