+++ docs/api-based-interface.ts
@@ -1,0 +1,8 @@
+import { ThirdParty } from '../interface';
+
+export type uuid = string | number;
+
+export interface ApiBasedLoginData {
+  thirdPartyId: ThirdParty;
+  uuid: uuid;
+}
